Tratando exceções globais com @ExceptionHandler e @ControllerAdvice e feito pequenas correções
